fix(import): make to create ErrorAlert component and use errorMessage spelling
